Tweaks to deal with unserialisation hiccups
On unserialising, things in Collections end up having their hashCode/equals
methods called before everything is set up, which might lead to NPEs with null
fields, so we deal with this. On the other hand, this produces invalid hashes,
so things unserialised into a HashMap need to be rehashed manually after loading.